feat: add limit order flow with price fraction validation

+++ src/main/java/com/mariesto/simplestockapp/model/UserStockResponse.java
@@ -1,7 +1,5 @@
 package com.mariesto.simplestockapp.model;
 
-import java.math.BigDecimal;
-
-public record UserStockResponse(String stockSymbol, Long quantity, BigDecimal totalPrice, BigDecimal currentPrice) {
+public record UserStockResponse(String stockSymbol, Long quantity, Long totalPrice, Long currentPrice) {
 }
